Revise infinite loop timeout handler.

+++ Library/StdDriver/inc/hdiv.h
@@ -1,8 +1,8 @@
 /**************************************************************************//**
  * @file     hdiv.h
  * @version  V3.0
- * $Revision: 4 $
- * $Date: 16/10/25 4:25p $
+ * $Revision: 1 $
+ * $Date: 16/07/07 7:50p $
  * @brief    NUC126 series Hardware Divider(HDIV) driver header file
  *
  * @note
@@ -13,8 +13,6 @@
  ******************************************************************************/
 #ifndef __HDIV_H__
 #define __HDIV_H__
-
-#include "NUC126.h"
 
 #ifdef __cplusplus
 extern "C"
@@ -50,9 +48,9 @@ static __INLINE int32_t HDIV_Div(int32_t x, int16_t y)
     uint32_t *p32;
 
     p32 = (uint32_t *)HDIV_BASE;
-    *p32++ = x;
-    *p32++ = y;
-    return *p32;
+    *p32++ = (uint32_t)x;
+    *p32++ = (uint32_t)y;
+    return (int32_t) * p32;
 }
 
 
@@ -71,9 +69,9 @@ static __INLINE int16_t HDIV_Mod(int32_t x, int16_t y)
     uint32_t *p32;
 
     p32 = (uint32_t *)HDIV_BASE;
-    *p32++ = x;
-    *p32++ = y;
-    return p32[1];
+    *p32++ = (uint32_t)x;
+    *p32++ = (uint32_t)y;
+    return (int16_t)p32[1];
 }
 
 /*@}*/ /* end of group HDIV_EXPORTED_FUNCTIONS */
@@ -89,5 +87,3 @@ static __INLINE int16_t HDIV_Mod(int32_t x, int16_t y)
 #endif //__HDIV_H__
 
 /*** (C) COPYRIGHT 2016 Nuvoton Technology Corp. ***/
-
-
